+++ client/home/TempWatchList.jsx
@@ -2,7 +2,6 @@ import React, { useState } from 'react';
 import { Text, View, StyleSheet } from 'react-native';
 import LinearView from '../sharedComponents/LinearView.jsx';
 
-import AddLeague from '../wishList/AddLeague.jsx';
 import LeaguesSaved from '../wishList/LeaguesSaved.jsx';
 
 import { mockData } from '../sharedComponents/mockData.js';
@@ -12,7 +11,7 @@ export default function TempWatchList({ navigation }) {
   const [wishList, setWishList] = useState(mockData.leagues);
   return (
     <LinearView>
-      {JSON.stringify(wishList) === '{}' ? <AddLeague navigation={navigation} setWishList={setWishList} /> : <LeaguesSaved navigation={navigation} wishList={wishList}/>}
+      <LeaguesSaved navigation={navigation} wishList={wishList} />
     </LinearView>
   )
 }
